requirements.txt
absl-py==0.7.0
appnope==0.1.0
astor==0.7.1
astroid==2.1.0
autopep8==1.4.3
backcall==0.1.0
cycler==0.10.0
decorator==4.3.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
matplotlib==3.0.2
mccabe==0.6.1
mock==2.0.0
numpy==1.16.1
parso==0.3.4
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.8
protobuf==3.6.1
ptyprocess==0.6.0
pycodestyle==2.4.0
Pygments==2.3.1
pylint==2.2.2
pyparsing==2.3.1
python-dateutil==2.8.0
six==1.11.0
tensorboard==1.12.2
tensorflow==1.13.0rc1
tensorflow-estimator==1.13.0rc0
termcolor==1.1.0
traitlets==4.3.2
virtualenv==16.1.0
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
